# docs/swagger.yaml
basePath: /api/v6
consumes:
- application/json
definitions:
  choices.NotificationChoice:
    enum:
    - REACTION
    - COMMENT
    - REPLY
    - ADMIN
    type: string
    x-enum-varnames:
    - NREACTION
    - NCOMMENT
    - NREPLY
    - NADMIN
  models.City:
    properties:
      country:
        example: Nigeria
        type: string
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      name:
        example: Lekki
        type: string
      region:
        example: Lagos
        type: string
      updated_at:
        type: string
    type: object
  models.Comment:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorID:
        type: string
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      post:
        $ref: '#/definitions/models.Post'
      postID:
        type: string
      slug:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  models.File:
    properties:
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      resource_type:
        type: string
      updated_at:
        type: string
    type: object
  models.Notification:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      commentID:
        type: string
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      ntype:
        $ref: '#/definitions/choices.NotificationChoice'
      post:
        $ref: '#/definitions/models.Post'
      postID:
        type: string
      readBy:
        items:
          $ref: '#/definitions/models.User'
        type: array
      receivers:
        items:
          $ref: '#/definitions/models.User'
        type: array
      reply:
        $ref: '#/definitions/models.Reply'
      replyID:
        type: string
      sender:
        $ref: '#/definitions/models.User'
      senderID:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  models.Post:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorID:
        type: string
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      image:
        $ref: '#/definitions/models.File'
      imageID:
        type: string
      slug:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  models.Reply:
    properties:
      author:
        $ref: '#/definitions/models.User'
      authorID:
        type: string
      comment:
        $ref: '#/definitions/models.Comment'
      commentID:
        type: string
      created_at:
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      slug:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      avatar:
        example: https://img.com
        type: string
      bio:
        example: Software Engineer | Go Fiber Developer
        type: string
      city:
        example: Lekki
        type: string
      created_at:
        type: string
      dob:
        type: string
      email:
        example: johndoe@email.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      last_name:
        example: Doe
        type: string
      updated_at:
        type: string
      username:
        example: john-doe
        type: string
    type: object
...
  schemas.AcceptFriendRequestSchema:
    properties:
      accepted:
        example: true
        type: boolean
      username:
        example: john-doe
        type: string
    required:
    - username
    type: object
  schemas.CitiesResponseSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/models.City'
        type: array
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.DeleteUserSchema:
    properties:
      password:
        example: password
        type: string
    required:
    - password
    type: object
...
  schemas.NotificationsResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      notifications:
        items:
          $ref: '#/definitions/models.Notification'
        type: array
      per_page:
        example: 100
        type: integer
    type: object
  schemas.NotificationsResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.NotificationsResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ProfileResponseSchema:
    properties:
      data:
        $ref: '#/definitions/models.User'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ProfileUpdateSchema:
    properties:
      bio:
        example: Software Engineer | Go Fiber Developer
        maxLength: 200
        type: string
      city_id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      dob:
        example: "2001-01-16T00:00:00.106416+01:00"
        type: string
      file_type:
        example: image/jpeg
        type: string
      first_name:
        example: John
        maxLength: 50
        minLength: 1
        type: string
      last_name:
        example: Doe
        maxLength: 50
        minLength: 1
        type: string
    type: object
  schemas.ProfilesResponseDataSchema:
    properties:
      current_page:
        example: 1
        type: integer
      last_page:
        example: 100
        type: integer
      per_page:
        example: 100
        type: integer
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  schemas.ProfilesResponseSchema:
    properties:
      data:
        $ref: '#/definitions/schemas.ProfilesResponseDataSchema'
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.ReadNotificationSchema:
    properties:
      id:
        example: d10dde64-a242-4ed0-bd75-4c759644b3a6
        type: string
      mark_all_as_read:
        example: false
        type: boolean
    type: object
...
  schemas.ResponseSchema:
    properties:
      message:
        example: Data fetched/created/updated/deleted
        type: string
      status:
        example: success
        type: string
    type: object
  schemas.SendFriendRequestSchema:
    properties:
      username:
        example: john-doe
        type: string
    required:
    - username
    type: object
...
info:
  contact: {}
  description: |
    ## A Realtime Social Networking API built with FIBER & GORM ORM.

    ### WEBSOCKETS:

    #### Notifications

    - URL: `wss://{host}/api/v6/ws/notifications`

    - Requires authorization, so pass in the Bearer Authorization header.

    - You can only read and not send notification messages into this socket.


    #### Chats

    - URL: `wss://{host}/api/v6/ws/chats/{id}`
    - Requires authorization, so pass in the Bearer Authorization header.
    - Use chat_id as the ID for an existing chat or username if it's the first message in a DM.
    - You cannot read realtime messages from a username that doesn't belong to the authorized user, but you can surely send messages.
    - Only send a message to the socket endpoint after the message has been created or updated, and files have been uploaded.
    - Fields when sending a message through the socket:

      ```json
      { "status": "CREATED", "id": "fe4e0235-80fc-4c94-b15e-3da63226f8ab" }
      ```
  title: SOCIALNET API
  version: "6.0"
paths:
...
  /profiles:
    get:
      description: This endpoint retrieves a paginated list of users
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfilesResponseSchema'
      security:
      - BearerAuth: []
      summary: Retrieve Users
      tags:
      - Profiles
  /profiles/cities:
    get:
      description: This endpoint retrieves the first 10 cities that matches the query
        params
      parameters:
      - description: City name
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.CitiesResponseSchema'
      summary: Retrieve cities based on query params
      tags:
      - Profiles
  /profiles/friends:
    get:
      description: This endpoint retrieves friends of a user
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfilesResponseSchema'
      security:
      - BearerAuth: []
      summary: Retrieve Friends
      tags:
      - Profiles
  /profiles/friends/requests:
    get:
      description: This endpoint retrieves friend requests of a user
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfilesResponseSchema'
      security:
      - BearerAuth: []
      summary: Retrieve Friend Requests
      tags:
      - Profiles
    post:
      description: This endpoint sends or delete friend requests
      parameters:
      - description: Friend Request object
        in: body
        name: friend_request
        required: true
        schema:
          $ref: '#/definitions/schemas.SendFriendRequestSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
      security:
      - BearerAuth: []
      summary: Send Or Delete Friend Request
      tags:
      - Profiles
    put:
      description: This endpoint accepts or reject a friend request
      parameters:
      - description: Friend Request object
        in: body
        name: friend_request
        required: true
        schema:
          $ref: '#/definitions/schemas.AcceptFriendRequestSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
      security:
      - BearerAuth: []
      summary: Accept Or Reject a Friend Request
      tags:
      - Profiles
  /profiles/notifications:
    get:
      description: This endpoint retrieves a paginated list of auth user's notifications.
        Use post, comment, reply slug to navigate to the post, comment or reply.
      parameters:
      - default: 1
        description: Current Page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.NotificationsResponseSchema'
      security:
      - BearerAuth: []
      summary: Retrieve User Notifications
      tags:
      - Profiles
    post:
      description: This endpoint reads a notification
      parameters:
      - description: Read Notification Data
        in: body
        name: read_data
        required: true
        schema:
          $ref: '#/definitions/schemas.ReadNotificationSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
      security:
      - BearerAuth: []
      summary: Read Notifications
      tags:
      - Profiles
  /profiles/profile:
    patch:
      description: This endpoint updates a user profile
      parameters:
      - description: Profile object
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/schemas.ProfileUpdateSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfileResponseSchema'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - Profiles
    post:
      description: This endpoint deletes a particular user's account (irreversible)
      parameters:
      - description: Password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/schemas.DeleteUserSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ResponseSchema'
      security:
      - BearerAuth: []
      summary: Delete User's Account
      tags:
      - Profiles
  /profiles/profile/{username}:
    get:
      description: This endpoint retrieves a user profile
      parameters:
      - description: Username of user
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.ProfileResponseSchema'
      summary: Retrieve User Profile
      tags:
      - Profiles
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer jwt_string' to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"